+++ src/map/packets/status_effects.cpp
@@ -43,8 +43,9 @@ CStatusEffectPacket::CStatusEffectPacket(CCharEntity* PChar)
         if (PEffect->GetIcon() != 0)
         {
             ref<uint16>(0x08 + (i * 0x02)) = PEffect->GetIcon();
+            // this value overflows, but the client expects the overflowed timestamp and corrects it
             ref<uint32>(0x48 + (i * 0x04)) = PEffect->GetDuration() == 0 ? 0x7FFFFFFF : 
-                (((PEffect->GetDuration() - std::chrono::duration_cast<std::chrono::milliseconds>(server_clock::now() - PEffect->GetStartTime()).count())/1000) 
+                (((PEffect->GetDuration() - (uint32)std::chrono::duration_cast<std::chrono::milliseconds>(server_clock::now() - PEffect->GetStartTime()).count())/1000) 
                 + CVanaTime::getInstance()->getVanaTime()) * 60;
             ++i;
         }
